whisper-server/requirements.txt
setuptools
wheel
fastapi
uvicorn
python-multipart
openai-whisper
torch
